feat/lazy-loading: add lazy-loading for variables page
