+++ src/app/[locale]/variable/page.tsx
@@ -1,34 +1,20 @@
-import { VariableEditor } from '@/components/variable-editor/variable-editor';
-import { Container, Typography } from '@mui/material';
-import { useTranslations } from 'next-intl';
+'use client'
+
+import { redirect } from '@/i18n/navigation';
+import { userAuthStore } from '@/store/userAuth/userAuth-store';
+import { useLocale } from 'next-intl';
+import dynamic from 'next/dynamic';
+
+const VariablePage = dynamic(() => import('../../../components/pages/variable-page'),   {
+  loading: () => <p>Loading...</p>,
+});
 
 export default function Page() {
-  const translate = useTranslations('VariablesPage');
+  const locale = useLocale();
+  const username = userAuthStore(state => state.userName);
 
-  return (
-    <Container
-      maxWidth="sm"
-      sx={{
-        mt: 4,
-        padding: '24px',
-        borderRadius: '8px',
-        boxShadow: '0px 4px 10px rgba(0, 0, 0, 0.1)',
-        backgroundColor: 'background.paper',
-      }}
-    >
-      <Typography
-        variant="h1"
-        sx={{ mb: 1, textAlign: 'center', fontSize: 40 }}
-      >
-        {translate('title')}
-      </Typography>
-      <Typography
-        variant="body1"
-        sx={{ mb: 2, textAlign: 'center', fontSize: 20 }}
-      >
-        {translate('description')}
-      </Typography>
-      <VariableEditor />
-    </Container>
-  );
+  return username ? <VariablePage /> : redirect({
+    href: '/',
+    locale
+  })
 }
